Move seed controller lifecycle into shared dir
